+++ graphqshell/src/application/mod.rs
@@ -1,0 +1,7 @@
+/// The application layer (in onion architecture)
+///
+/// The user or any other program interacts with the application(s) in this module.
+/// The module contain things like user interfaces, APIs, and provides orchestration for the domain
+/// logic. It might contain services and logic likewise.
+pub mod configuration;
+pub mod tui_app;
