+++ graphqshell/src/application/tui_app/tui/components/log.rs
