+++ graphqshell/src/domain/mod.rs
@@ -1,0 +1,6 @@
+/// The domain layer (in onion architecture)
+///
+/// This contains all the logic and data that belongs to the domain model.
+/// So in this case it is all about GraphQL schema analysis and interaction with a
+/// GraphQL API
+pub mod graphql;
