+++ graphqshell/src/infra/mod.rs
@@ -1,0 +1,5 @@
+/// The infrastructure layer (in onion architecture)
+///
+/// This module holds the implementations that provide access to required external components.
+/// It essentially provides concrete implementations for the contracts used domain and application.
+pub mod graphql_api;
