+++ graphqshell/src/lib.rs
@@ -5,7 +5,7 @@ extern crate tui;
 #[cfg(test)]
 #[macro_use]
 extern crate matches;
-pub mod adapter;
+pub mod application;
 pub mod domain;
-pub mod use_case;
+pub mod infra;
 pub mod utils;
