--- graphqshell/src/use_case/mod.rs
@@ -1,1 +1,0 @@
-
